+++ src/services/index.js
@@ -1,5 +1,5 @@
 import authService from './auth-service';
 import postsService from './posts-service';
-import tokenHandler from './token-hander';
+import tokenHandler from './token-handler';
 
 export { authService, postsService, tokenHandler };
